Now job_from_file is invoked by the right user and not by the db server
